// anarun/types/corrida.Type.ts
export type CorridaType = {
  id: number
  distancia: number
  tempo: number
  bpm: number
  pace: string
  data: string
}

export type DiarioType = {
  id: number
  data: string
  sono: number
  hidratacao: number
  diasDescanso: number
  alimentacao: string
  dorMuscular: string
  temperatura: number
  indice: number
}